extract ensure ConfigMaps and Secrets
